+++ NEWS.md
@@ -3,7 +3,7 @@
 * Expanded framework for tidy radiocarbon analysis, adding:
   * New S3 object `cal`, a generic representation of calibrated probability distributions
   * Methods for printing and summarising `cal` objects
-  * Type conversion functions for `cal` objects
+  * Type conversion functions for `cal` objects (`as_cal.CalDates()`, `as_cal.oxcAARCalibratedDatesList()` and `as_cal.BchronCalibratedDates()`)
   * New `vignette("tidy_radiocarbon")`
 * **Breaking change**: `cal()` and `sum()` renamed `c14_calibrate` and `c14_sum` and now return `cal` objects.
 * **Breaking changes** to `stratigraph` interface: